@@ -14,7 +14,7 @@
     }
   },
   "require": {
-    "php": ">=5.4.9",
+    "php": ">=5.5.9",
     "oro/commerce": "self.version",
     "nelmio/alice": "^2.1"
   },
