@@ -15,22 +15,28 @@ use Oro\Bundle\ImportExportBundle\Field\DatabaseHelper;
 
 class ConfigurableAddOrReplaceStrategy extends AbstractImportStrategy
 {
+    const STRATEGY_CONTEXT = 'configurable_add_or_replace_strategy';
+
     /** @var ChainEntityClassNameProvider */
     protected $chainEntityClassNameProvider;
 
     /** @var TranslatorInterface */
     protected $translator;
 
+    /** @var NewEntitiesHelper */
+    protected $newEntitiesHelper;
+
     /** @var array */
-    protected $cachedEntities = array();
+    protected $cachedEntities = [];
 
     /**
-     * @param EventDispatcherInterface $eventDispatcher
-     * @param ImportStrategyHelper $strategyHelper
-     * @param FieldHelper $fieldHelper
-     * @param DatabaseHelper $databaseHelper
+     * @param EventDispatcherInterface     $eventDispatcher
+     * @param ImportStrategyHelper         $strategyHelper
+     * @param FieldHelper                  $fieldHelper
+     * @param DatabaseHelper               $databaseHelper
      * @param ChainEntityClassNameProvider $chainEntityClassNameProvider
-     * @param TranslatorInterface $translator
+     * @param TranslatorInterface          $translator
+     * @param NewEntitiesHelper            $newEntitiesHelper
      */
     public function __construct(
         EventDispatcherInterface $eventDispatcher,
@@ -38,11 +44,13 @@ class ConfigurableAddOrReplaceStrategy extends AbstractImportStrategy
         FieldHelper $fieldHelper,
         DatabaseHelper $databaseHelper,
         ChainEntityClassNameProvider $chainEntityClassNameProvider,
-        TranslatorInterface $translator
+        TranslatorInterface $translator,
+        NewEntitiesHelper $newEntitiesHelper
     ) {
         parent::__construct($eventDispatcher, $strategyHelper, $fieldHelper, $databaseHelper);
         $this->chainEntityClassNameProvider = $chainEntityClassNameProvider;
-        $this->translator = $translator;
+        $this->translator                   = $translator;
+        $this->newEntitiesHelper            = $newEntitiesHelper;
     }
 
 
@@ -53,7 +61,7 @@ class ConfigurableAddOrReplaceStrategy extends AbstractImportStrategy
     {
         $this->assertEnvironment($entity);
 
-        $this->cachedEntities = array();
+        $this->cachedEntities = [];
 
         if (!$entity = $this->beforeProcessEntity($entity)) {
             return null;
@@ -71,12 +79,12 @@ class ConfigurableAddOrReplaceStrategy extends AbstractImportStrategy
     }
 
     /**
-     * @param object $entity
-     * @param bool   $isFullData
-     * @param bool   $isPersistNew
+     * @param object           $entity
+     * @param bool             $isFullData
+     * @param bool             $isPersistNew
      * @param mixed|array|null $itemData
-     * @param array $searchContext
-     * @param bool $entityIsRelation
+     * @param array            $searchContext
+     * @param bool             $entityIsRelation
      *
      * @return null|object
      */
@@ -85,14 +93,14 @@ class ConfigurableAddOrReplaceStrategy extends AbstractImportStrategy
         $isFullData = false,
         $isPersistNew = false,
         $itemData = null,
-        array $searchContext = array(),
+        array $searchContext = [],
         $entityIsRelation = false
     ) {
         $oid = spl_object_hash($entity);
         if (isset($this->cachedEntities[$oid])) {
             return $entity;
         }
-
+        $entityClass = ClassUtils::getClass($entity);
         // find and cache existing or new entity
         $existingEntity = $this->findExistingEntity($entity, $searchContext);
         if ($existingEntity) {
@@ -105,7 +113,7 @@ class ConfigurableAddOrReplaceStrategy extends AbstractImportStrategy
             // if can't find entity and new entity can't be persisted
             if (!$isPersistNew) {
                 if ($entityIsRelation) {
-                    $class = $this->chainEntityClassNameProvider->getEntityClassName(ClassUtils::getClass($entity));
+                    $class         = $this->chainEntityClassNameProvider->getEntityClassName($entityClass);
                     $errorMessages = [$this->translator->trans(
                         'oro.importexport.import.errors.not_found_entity',
                         ['%entity_name%' => $class]
@@ -114,6 +122,24 @@ class ConfigurableAddOrReplaceStrategy extends AbstractImportStrategy
                 }
 
                 return null;
+            } else {
+                /**
+                 * Save new entity to newEntitiesHelper storage by key constructed from identityValues
+                 * and this strategy context for reuse if there will be entity with the same identity values
+                 * it has not be created again but has be fetch from this storage
+                 */
+                $identityValues = $this->combineIdentityValues($entity, $entityClass, $searchContext);
+                if ($identityValues) {
+                    $newEntityKey   = sprintf('%s:%s', $entityClass, serialize($identityValues));
+                    $existingEntity = $this->newEntitiesHelper->getEntity($newEntityKey);
+                    if (null === $existingEntity) {
+                        $this->newEntitiesHelper->setEntity($newEntityKey, $entity);
+                        $this->newEntitiesHelper->incrementEntityUsage($this->getEntityHashKey($entity));
+                    } else {
+                        $this->newEntitiesHelper->incrementEntityUsage($this->getEntityHashKey($existingEntity));
+                    }
+                }
+
             }
 
             $this->databaseHelper->resetIdentifier($entity);
@@ -138,21 +164,21 @@ class ConfigurableAddOrReplaceStrategy extends AbstractImportStrategy
     }
 
     /**
-     * @param object $entity
-     * @param object $existingEntity
+     * @param object           $entity
+     * @param object           $existingEntity
      * @param mixed|array|null $itemData
-     * @param array $excludedFields
+     * @param array            $excludedFields
      */
     protected function importExistingEntity(
         $entity,
         $existingEntity,
         $itemData = null,
-        array $excludedFields = array()
+        array $excludedFields = []
     ) {
-        $entityName = ClassUtils::getClass($entity);
-        $identifierName = $this->databaseHelper->getIdentifierFieldName($entityName);
+        $entityName       = ClassUtils::getClass($entity);
+        $identifierName   = $this->databaseHelper->getIdentifierFieldName($entityName);
         $excludedFields[] = $identifierName;
-        $fields = $this->fieldHelper->getFields($entityName, true);
+        $fields           = $this->fieldHelper->getFields($entityName, true);
 
         foreach ($fields as $key => $field) {
             $fieldName = $field['name'];
@@ -168,9 +194,10 @@ class ConfigurableAddOrReplaceStrategy extends AbstractImportStrategy
     /**
      * Exclude fields marked as "excluded" and skipped not identity fields
      *
-     * @param string $entityName
-     * @param string $fieldName
+     * @param string           $entityName
+     * @param string           $fieldName
      * @param array|mixed|null $itemData
+     *
      * @return bool
      */
     protected function isFieldExcluded($entityName, $fieldName, $itemData = null)
@@ -183,23 +210,23 @@ class ConfigurableAddOrReplaceStrategy extends AbstractImportStrategy
     }
 
     /**
-     * @param object $entity
+     * @param object     $entity
      * @param array|null $itemData
      */
     protected function updateRelations($entity, array $itemData = null)
     {
         $entityName = ClassUtils::getClass($entity);
-        $fields = $this->fieldHelper->getFields($entityName, true);
+        $fields     = $this->fieldHelper->getFields($entityName, true);
 
         foreach ($fields as $field) {
             if ($this->fieldHelper->isRelation($field)) {
-                $fieldName = $field['name'];
-                $isFullRelation = $this->fieldHelper->getConfigValue($entityName, $fieldName, 'full', false);
+                $fieldName         = $field['name'];
+                $isFullRelation    = $this->fieldHelper->getConfigValue($entityName, $fieldName, 'full', false);
                 $isPersistRelation = $this->databaseHelper->isCascadePersist($entityName, $fieldName);
                 $inversedFieldName = $this->databaseHelper->getInversedRelationFieldName($entityName, $fieldName);
 
                 // additional search parameters to find only related entities
-                $searchContext = array();
+                $searchContext = [];
                 if ($isPersistRelation && $inversedFieldName
                     && $this->databaseHelper->isSingleInversedRelation($entityName, $fieldName)
                 ) {
@@ -211,7 +238,7 @@ class ConfigurableAddOrReplaceStrategy extends AbstractImportStrategy
                     $relationEntity = $this->fieldHelper->getObjectValue($entity, $fieldName);
                     if ($relationEntity) {
                         $relationItemData = $this->fieldHelper->getItemData($itemData, $fieldName);
-                        $relationEntity = $this->processEntity(
+                        $relationEntity   = $this->processEntity(
                             $relationEntity,
                             $isFullRelation,
                             $isPersistRelation,
@@ -229,7 +256,7 @@ class ConfigurableAddOrReplaceStrategy extends AbstractImportStrategy
                         $collectionEntities = new ArrayCollection();
 
                         foreach ($relationCollection as $collectionEntity) {
-                            $entityItemData = $this->fieldHelper->getItemData(array_shift($collectionItemData));
+                            $entityItemData   = $this->fieldHelper->getItemData(array_shift($collectionItemData));
                             $collectionEntity = $this->processEntity(
                                 $collectionEntity,
                                 $isFullRelation,
@@ -253,6 +280,7 @@ class ConfigurableAddOrReplaceStrategy extends AbstractImportStrategy
 
     /**
      * @param object $entity
+     *
      * @return null|object
      */
     protected function validateAndUpdateContext($entity)
@@ -279,10 +307,60 @@ class ConfigurableAddOrReplaceStrategy extends AbstractImportStrategy
     protected function updateContextCounters($entity)
     {
         $identifier = $this->databaseHelper->getIdentifier($entity);
-        if ($identifier) {
+        if ($identifier || $this->newEntitiesHelper->getEntityUsage($this->getEntityHashKey($entity)) > 1) {
             $this->context->incrementReplaceCount();
         } else {
             $this->context->incrementAddCount();
         }
+    }
+
+    /**
+     * @param object $entity
+     *
+     * @return string
+     */
+    protected function getEntityHashKey($entity)
+    {
+        $hashKey = self::STRATEGY_CONTEXT . spl_object_hash($entity);
+
+        return $hashKey;
+    }
+
+    /**
+     * Combines identity values for entity search on local new entities storage
+     * (which are not yet saved in db)
+     * from search context and not empty identity fields or required identity fields
+     * which could be null if configured.
+     * At least one not null and not empty value must be present for search
+     *
+     * @param       $entity
+     * @param       $entityClass
+     * @param array $searchContext
+     *
+     * @return array|null
+     */
+    protected function combineIdentityValues($entity, $entityClass, array $searchContext)
+    {
+        $identityValues = $searchContext;
+        $identityValues += $this->fieldHelper->getIdentityValues($entity);
+        $notEmptyValues     = [];
+        $nullRequiredValues = [];
+        foreach ($identityValues as $fieldName => $value) {
+            if (null !== $value) {
+                if ('' !== $value) {
+                    if (is_object($value)) {
+                        $notEmptyValues[$fieldName] = $this->databaseHelper->getIdentifier($value);
+                    } else {
+                        $notEmptyValues[$fieldName] = $value;
+                    }
+                }
+            } elseif ($this->fieldHelper->isRequiredIdentityField($entityClass, $fieldName)) {
+                $nullRequiredValues[$fieldName] = null;
+            }
+        }
+
+        return !empty($notEmptyValues)
+            ? array_merge($notEmptyValues, $nullRequiredValues)
+            : null;
     }
 }
