@@ -9,15 +9,19 @@ use OroB2B\Bundle\CMSBundle\Entity\Page;
 
 class LoadPageData extends AbstractFixture
 {
+    const PAGE_1 = 'page.1';
+    const PAGE_1_2 = 'page.1_2';
+    const PAGE_1_3 = 'page.1_3';
+
     /**
      * @var array
      */
     protected static $page = [
-        'page.1' => [],
-        'page.1_2' => [
+        self::PAGE_1 => [],
+        self::PAGE_1_2 => [
             'parent' => 'page.1'
         ],
-        'page.1_3' => [
+        self::PAGE_1_3 => [
             'parent' => 'page.1'
         ],
     ];
