@@ -28,7 +28,7 @@ abstract class AbstractCategoryFixture extends AbstractFixture
         $categoryRepository = $manager->getRepository('OroB2BCatalogBundle:Category');
         $root = $categoryRepository->getMasterCatalogRoot();
 
-        $this->addCategories($root, $this->categories);
+        $this->addCategories($root, $this->categories, $manager);
 
         $manager->flush();
     }
@@ -37,7 +37,7 @@ abstract class AbstractCategoryFixture extends AbstractFixture
      * @param Category $root
      * @param array $categories
      */
-    protected function addCategories(Category $root, array $categories)
+    protected function addCategories(Category $root, array $categories, ObjectManager $manager)
     {
         if (!$categories) {
             return;
@@ -50,11 +50,13 @@ abstract class AbstractCategoryFixture extends AbstractFixture
             $category = new Category();
             $category->addTitle($categoryTitle);
 
+            $manager->persist($category);
+
             $this->addReference($title, $category);
 
             $root->addChildCategory($category);
 
-            $this->addCategories($category, $nestedCategories);
+            $this->addCategories($category, $nestedCategories, $manager);
         }
     }
 }
