@@ -31,8 +31,15 @@ class PaymentMethodExtensionTest extends \PHPUnit_Framework_TestCase
      */
     protected $extension;
 
+    /**
+     * @var \PHPUnit_Framework_MockObject_MockObject|TranslatorInterface
+     */
+    protected $translator;
+
+
     public function setUp()
     {
+        $this->translator = $this->getMock('Symfony\Component\Translation\TranslatorInterface');
         $this->paymentTransactionProvider = $this
             ->getMockBuilder('OroB2B\Bundle\PaymentBundle\Provider\PaymentTransactionProvider')
             ->disableOriginalConstructor()
@@ -43,7 +50,8 @@ class PaymentMethodExtensionTest extends \PHPUnit_Framework_TestCase
             ->getMock();
         $this->extension = new PaymentMethodExtension(
             $this->paymentTransactionProvider,
-            $this->paymentMethodViewRegistry
+            $this->paymentMethodViewRegistry,
+            $this->translator
         );
         $this->paymentMethodView = $this
             ->getMockBuilder('OroB2B\Bundle\PaymentBundle\Method\View\PaymentMethodViewInterface')
@@ -56,7 +64,11 @@ class PaymentMethodExtensionTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals(
             [
                 new \Twig_SimpleFunction('get_payment_methods', [$this->extension, 'getPaymentMethods']),
-                new \Twig_SimpleFunction('get_payment_method_label', [$this->extension, 'getPaymentMethodLabel'])
+                new \Twig_SimpleFunction('get_payment_method_label', [$this->extension, 'getPaymentMethodLabel']),
+                new \Twig_SimpleFunction(
+                    'get_payment_method_admin_label',
+                    [$this->extension, 'getPaymentMethodAdminLabel']
+                )
             ],
             $this->extension->getFunctions()
         );
@@ -65,6 +77,24 @@ class PaymentMethodExtensionTest extends \PHPUnit_Framework_TestCase
     public function testGetName()
     {
         $this->assertEquals(PaymentMethodExtension::PAYMENT_METHOD_EXTENSION_NAME, $this->extension->getName());
+    }
+
+    /**
+     * @param string $paymentMethod
+     * @param string $returnLabel
+     * @param bool $isShort
+     */
+    public function paymentMethodLabelMock($paymentMethod, $returnLabel, $isShort = true)
+    {
+        $this->paymentMethodViewRegistry
+            ->expects($this->once())
+            ->method('getPaymentMethodView')
+            ->with($paymentMethod)
+            ->willReturn($this->paymentMethodView);
+        $this->paymentMethodView
+            ->expects($this->once())
+            ->method($isShort?'getShortLabel':'getLabel')
+            ->willReturn($returnLabel);
     }
 
     public function testGetPaymentMethods()
@@ -79,15 +109,7 @@ class PaymentMethodExtensionTest extends \PHPUnit_Framework_TestCase
             ->method('getPaymentTransactions')
             ->with($entity)
             ->willReturn([$paymentTransaction]);
-        $this->paymentMethodViewRegistry
-            ->expects($this->once())
-            ->method('getPaymentMethodView')
-            ->with($paymentMethodConstant)
-            ->willReturn($this->paymentMethodView);
-        $this->paymentMethodView
-            ->expects($this->once())
-            ->method('getLabel')
-            ->willReturn($label);
+        $this->paymentMethodLabelMock($paymentMethodConstant, $label, false);
 
         $this->assertEquals($this->extension->getPaymentMethods($entity), [$label]);
     }
@@ -126,5 +148,50 @@ class PaymentMethodExtensionTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals($this->extension->getPaymentMethodLabel($paymentMethodConstant), $label);
         $this->assertEquals($this->extension->getPaymentMethodLabel($paymentMethodNotExistsConstant), '');
         $this->assertEquals($this->extension->getPaymentMethodLabel($paymentMethodConstant, false), $label);
+    }
+
+    /**
+     * @dataProvider paymentProvider
+     * @param string $paymentMethod
+     * @param string $paymentMethodLabel
+     * @param string $paymentMethodShortLabel
+     * @param string $expectedResult
+     */
+    public function testGetPaymentMethodAdminLabel(
+        $paymentMethod,
+        $paymentMethodLabel,
+        $paymentMethodShortLabel,
+        $expectedResult
+    ) {
+        $this->translator
+            ->expects($this->once())
+            ->method('trans')
+            ->with('orob2b.payment.admin.'.$paymentMethod.'.label')
+            ->willReturn($paymentMethodLabel);
+
+        $this->paymentMethodLabelMock($paymentMethod, $paymentMethodShortLabel);
+
+        $this->assertEquals($this->extension->getPaymentMethodAdminLabel($paymentMethod), $expectedResult);
+    }
+
+    /**
+     * @return array
+     */
+    public function paymentProvider()
+    {
+        return [
+            [
+                '$paymentMethod' => 'payment_method',
+                '$paymentMethodLabel' => 'Payment Method',
+                '$paymentMethodShortLabel' => 'Payment Method Short',
+                '$expectedResult' => 'Payment Method (Payment Method Short)',
+            ],
+            [
+                '$paymentMethod' => 'payment_method',
+                '$paymentMethodLabel' => 'Payment Method',
+                '$paymentMethodShortLabel' => 'Payment Method',
+                '$expectedResult' => 'Payment Method',
+            ],
+        ];
     }
 }
