@@ -2,6 +2,8 @@
 
 namespace OroB2B\Bundle\PaymentBundle\Twig;
 
+use Symfony\Component\Translation\TranslatorInterface;
+
 use OroB2B\Bundle\PaymentBundle\Method\View\PaymentMethodViewInterface;
 use OroB2B\Bundle\PaymentBundle\Method\View\PaymentMethodViewRegistry;
 use OroB2B\Bundle\PaymentBundle\Provider\PaymentTransactionProvider;
@@ -10,22 +12,34 @@ class PaymentMethodExtension extends \Twig_Extension
 {
     const PAYMENT_METHOD_EXTENSION_NAME = 'orob2b_payment_method';
 
-    /** @var  PaymentTransactionProvider */
+    /**
+     * @var  PaymentTransactionProvider
+     */
     protected $paymentTransactionProvider;
 
-    /** @var  PaymentMethodViewRegistry */
+    /**
+     * @var  PaymentMethodViewRegistry
+     */
     protected $paymentMethodViewRegistry;
+
+    /**
+     * @var TranslatorInterface
+     */
+    protected $translator;
 
     /**
      * @param PaymentTransactionProvider $paymentTransactionProvider
      * @param PaymentMethodViewRegistry $paymentMethodViewRegistry
+     * @param TranslatorInterface $translator
      */
     public function __construct(
         PaymentTransactionProvider $paymentTransactionProvider,
-        PaymentMethodViewRegistry $paymentMethodViewRegistry
+        PaymentMethodViewRegistry $paymentMethodViewRegistry,
+        TranslatorInterface $translator
     ) {
         $this->paymentTransactionProvider = $paymentTransactionProvider;
         $this->paymentMethodViewRegistry = $paymentMethodViewRegistry;
+        $this->translator = $translator;
     }
 
     /**
@@ -44,6 +58,7 @@ class PaymentMethodExtension extends \Twig_Extension
         return [
             new \Twig_SimpleFunction('get_payment_methods', [$this, 'getPaymentMethods']),
             new \Twig_SimpleFunction('get_payment_method_label', [$this, 'getPaymentMethodLabel']),
+            new \Twig_SimpleFunction('get_payment_method_admin_label', [$this, 'getPaymentMethodAdminLabel'])
         ];
     }
 
@@ -66,7 +81,7 @@ class PaymentMethodExtension extends \Twig_Extension
     /**
      * @param string $paymentMethod
      * @param bool $shortLabel
-     * @return array
+     * @return string
      */
     public function getPaymentMethodLabel($paymentMethod, $shortLabel = true)
     {
@@ -77,6 +92,22 @@ class PaymentMethodExtension extends \Twig_Extension
             return $shortLabel? $paymentMethodView->getShortLabel() : $paymentMethodView->getLabel();
         } catch (\InvalidArgumentException $e) {
             return '';
+        }
+    }
+
+    /**
+     * @param string $paymentMethod
+     * @return string
+     */
+    public function getPaymentMethodAdminLabel($paymentMethod)
+    {
+        $adminPaymentMethodLabel = $this->translator->trans('orob2b.payment.admin.'.$paymentMethod.'.label');
+        $adminPaymentMethodShortLabel = $this->getPaymentMethodLabel($paymentMethod);
+
+        if ($adminPaymentMethodLabel == $adminPaymentMethodShortLabel) {
+            return $adminPaymentMethodLabel;
+        } else {
+            return $adminPaymentMethodLabel.' ('.$adminPaymentMethodShortLabel.')';
         }
     }
 }
