@@ -33,7 +33,7 @@ class OroB2BPricingBundleInstaller implements Installation, NoteExtensionAwareIn
      */
     public function getMigrationVersion()
     {
-        return 'v1_3';
+        return 'v1_4';
     }
 
     /**
@@ -63,6 +63,9 @@ class OroB2BPricingBundleInstaller implements Installation, NoteExtensionAwareIn
         $this->createOroB2BProductPriceChangeTriggerTable($schema);
         $this->createOroB2BPriceListScheduleTable($schema);
         $this->createOroB2BCplActivationRuleTable($schema);
+        $this->createOroB2BPriceAttributeTable($schema);
+        $this->createOroB2BPriceAttributeCurrencyTable($schema);
+        $this->createOroB2BPriceAttributeProductPriceTable($schema);
 
         /** Foreign keys generation **/
         $this->addOrob2BPriceListCurrencyForeignKeys($schema);
@@ -84,6 +87,8 @@ class OroB2BPricingBundleInstaller implements Installation, NoteExtensionAwareIn
         $this->addOroB2BProductPriceChangeTriggerForeignKeys($schema);
         $this->addOroB2BPriceListScheduleForeignKeys($schema);
         $this->addOroB2BCplActivationRuleForeignKeys($schema);
+        $this->addOroB2BPriceAttributeCurrencyForeignKeys($schema);
+        $this->addOroB2BPriceAttributeProductPriceForeignKeys($schema);
     }
 
     /**
@@ -457,6 +462,53 @@ class OroB2BPricingBundleInstaller implements Installation, NoteExtensionAwareIn
         $table->addColumn('account_id', 'integer', ['notnull' => false]);
         $table->setPrimaryKey(['id']);
     }
+
+    /**
+     * @param Schema $schema
+     */
+    protected function createOroB2BPriceAttributeTable(Schema $schema)
+    {
+        $table = $schema->createTable('orob2b_price_attribute_pl');
+        $table->addColumn('id', 'integer', ['autoincrement' => true]);
+        $table->addColumn('name', 'string', ['length' => 255]);
+        $table->addColumn('created_at', 'datetime', ['comment' => '(DC2Type:datetime)']);
+        $table->addColumn('updated_at', 'datetime', ['comment' => '(DC2Type:datetime)']);
+        $table->setPrimaryKey(['id']);
+    }
+
+    /**
+     * @param Schema $schema
+     */
+    protected function createOroB2BPriceAttributeCurrencyTable(Schema $schema)
+    {
+        $table = $schema->createTable('orob2b_product_attr_currency');
+        $table->addColumn('id', 'integer', ['autoincrement' => true]);
+        $table->addColumn('price_attribute_pl_id', 'integer', []);
+        $table->addColumn('currency', 'string', ['length' => 3]);
+        $table->setPrimaryKey(['id']);
+    }
+
+    /**
+     * @param Schema $schema
+     */
+    protected function createOroB2BPriceAttributeProductPriceTable(Schema $schema)
+    {
+        $table = $schema->createTable('orob2b_price_attribute_price');
+        $table->addColumn('id', 'integer', ['autoincrement' => true]);
+        $table->addColumn('price_attribute_pl_id', 'integer', []);
+        $table->addColumn('product_id', 'integer', []);
+        $table->addColumn('unit_code', 'string', ['length' => 255]);
+        $table->addColumn('product_sku', 'string', ['length' => 255]);
+        $table->addColumn('quantity', 'float', []);
+        $table->addColumn('value', 'money', []);
+        $table->addColumn('currency', 'string', ['length' => 3]);
+        $table->addUniqueIndex(
+            ['product_id', 'price_attribute_pl_id', 'quantity', 'unit_code', 'currency'],
+            'orob2b_pricing_price_attribute_uidx'
+        );
+        $table->setPrimaryKey(['id']);
+    }
+
 
     /**
      * Add orob2b_price_list_currency foreign keys.
@@ -915,6 +967,46 @@ class OroB2BPricingBundleInstaller implements Installation, NoteExtensionAwareIn
             ['combined_price_list_id'],
             ['id'],
             ['onDelete' => 'CASCADE', 'onUpdate' => null]
+        );
+    }
+
+    /**
+     * @param Schema $schema
+     */
+    protected function addOroB2BPriceAttributeCurrencyForeignKeys(Schema $schema)
+    {
+        $table = $schema->getTable('orob2b_product_attr_currency');
+        $table->addForeignKeyConstraint(
+            $schema->getTable('orob2b_price_attribute_pl'),
+            ['price_attribute_pl_id'],
+            ['id'],
+            ['onUpdate' => null, 'onDelete' => 'CASCADE']
+        );
+    }
+
+    /**
+     * @param Schema $schema
+     */
+    protected function addOroB2BPriceAttributeProductPriceForeignKeys(Schema $schema)
+    {
+        $table = $schema->getTable('orob2b_price_attribute_price');
+        $table->addForeignKeyConstraint(
+            $schema->getTable('orob2b_price_attribute_pl'),
+            ['price_attribute_pl_id'],
+            ['id'],
+            ['onUpdate' => null, 'onDelete' => 'CASCADE']
+        );
+        $table->addForeignKeyConstraint(
+            $schema->getTable('orob2b_product'),
+            ['product_id'],
+            ['id'],
+            ['onUpdate' => null, 'onDelete' => 'CASCADE']
+        );
+        $table->addForeignKeyConstraint(
+            $schema->getTable('orob2b_product_unit'),
+            ['unit_code'],
+            ['code'],
+            ['onUpdate' => null, 'onDelete' => 'CASCADE']
         );
     }
 }
