@@ -53,7 +53,8 @@ class ProductPriceResetStrategyTest extends WebTestCase
             $container->get('oro_importexport.field.field_helper'),
             $container->get('oro_importexport.field.database_helper'),
             $container->get('oro_entity.entity_class_name_provider'),
-            $container->get('translator')
+            $container->get('translator'),
+            $container->get('oro_importexport.strategy.new_entities_helper')
         );
         
         $this->stepExecution = new StepExecution('step', new JobExecution());
