@@ -2,13 +2,13 @@
 
 namespace OroB2B\Bundle\ProductBundle\Migrations\Data\Demo\ORM;
 
-use Symfony\Component\Config\FileLocator;
-use Symfony\Component\DependencyInjection\ContainerAwareInterface;
-use Symfony\Component\DependencyInjection\ContainerInterface;
-
 use Doctrine\Common\DataFixtures\AbstractFixture;
 use Doctrine\Common\Persistence\ObjectManager;
 use Doctrine\ORM\EntityManager;
+
+use Symfony\Component\Config\FileLocator;
+use Symfony\Component\DependencyInjection\ContainerAwareInterface;
+use Symfony\Component\DependencyInjection\ContainerInterface;
 
 use Oro\Bundle\EntityExtendBundle\Entity\AbstractEnumValue;
 use Oro\Bundle\EntityExtendBundle\Tools\ExtendHelper;
@@ -16,6 +16,7 @@ use Oro\Bundle\LocaleBundle\Entity\LocalizedFallbackValue;
 use Oro\Bundle\UserBundle\DataFixtures\UserUtilityTrait;
 
 use OroB2B\Bundle\ProductBundle\Entity\Product;
+use OroB2B\Bundle\ProductBundle\Entity\ProductImage;
 use OroB2B\Bundle\ProductBundle\Entity\ProductUnit;
 use OroB2B\Bundle\ProductBundle\Entity\ProductUnitPrecision;
 
@@ -61,6 +62,8 @@ class LoadProductDemoData extends AbstractFixture implements ContainerAwareInter
         $headers = fgetcsv($handler, 1000, ',');
 
         $inventoryStatuses = $this->getAllEnumValuesByCode($manager, 'prod_inventory_status');
+
+        $allImageTypes = $this->getImageTypes();
 
         while (($data = fgetcsv($handler, 1000, ',')) !== false) {
             $row = array_combine($headers, array_values($data));
@@ -110,9 +113,9 @@ class LoadProductDemoData extends AbstractFixture implements ContainerAwareInter
             
             $product->setPrimaryUnitPrecision($productUnitPrecision);
 
-            $image = $this->getImageForProductSku($manager, $locator, $row['sku']);
-            if ($image) {
-                $product->setImage($image);
+            $productImage = $this->getProductImageForProductSku($manager, $locator, $row['sku'], $allImageTypes);
+            if ($productImage) {
+                $product->addImage($productImage);
             }
 
             $manager->persist($product);
@@ -139,11 +142,12 @@ class LoadProductDemoData extends AbstractFixture implements ContainerAwareInter
      * @param ObjectManager $manager
      * @param FileLocator $locator
      * @param string $sku
-     * @return null|\Oro\Bundle\AttachmentBundle\Entity\File
+     * @param array|null $types
+     * @return null|ProductImage
      */
-    protected function getImageForProductSku(ObjectManager $manager, FileLocator $locator, $sku)
+    protected function getProductImageForProductSku(ObjectManager $manager, FileLocator $locator, $sku, $types)
     {
-        $image = null;
+        $productImage = null;
 
         try {
             $imagePath = $locator->locate(sprintf('@OroB2BProductBundle/Migrations/Data/Demo/ORM/images/%s.jpg', $sku));
@@ -159,11 +163,27 @@ class LoadProductDemoData extends AbstractFixture implements ContainerAwareInter
             $attachmentManager->upload($image);
 
             $manager->persist($image);
+
+            $productImage = new ProductImage();
+            $productImage->setImage($image);
+            foreach ($types as $type) {
+                $productImage->addType($type);
+            }
         } catch (\Exception $e) {
             //image not found
         }
 
-        return $image;
+        return $productImage;
+    }
+
+    /**
+     * @return array
+     */
+    protected function getImageTypes()
+    {
+        $imageTypeProvider = $this->container->get('oro_layout.provider.image_type');
+
+        return array_keys($imageTypeProvider->getImageTypes());
     }
 
     /**
