@@ -32,6 +32,9 @@ class OroB2BProductBundleInstaller implements
     const MAX_PRODUCT_IMAGE_SIZE_IN_MB = 10;
     const MAX_PRODUCT_ATTACHMENT_SIZE_IN_MB = 5;
 
+    const PRODUCT_IMAGE_TABLE_NAME = 'orob2b_product_image';
+    const PRODUCT_IMAGE_TYPE_TABLE_NAME = 'orob2b_product_image_type';
+
     /** @var ExtendExtension */
     protected $extendExtension;
 
@@ -70,7 +73,7 @@ class OroB2BProductBundleInstaller implements
      */
     public function getMigrationVersion()
     {
-        return 'v1_5';
+        return 'v1_6';
     }
 
     /**
@@ -85,6 +88,8 @@ class OroB2BProductBundleInstaller implements
         $this->createOrob2BProductDescriptionTable($schema);
         $this->createOroB2BProductVariantLinkTable($schema);
         $this->createOroB2BProductShortDescriptionTable($schema);
+        $this->createOroB2BProductImageTable($schema);
+        $this->createOroB2BProductImageTypeTable($schema);
 
         $this->addOroB2BProductForeignKeys($schema);
         $this->addOroB2BProductUnitPrecisionForeignKeys($schema);
@@ -92,6 +97,8 @@ class OroB2BProductBundleInstaller implements
         $this->addOrob2BProductDescriptionForeignKeys($schema);
         $this->addOroB2BProductVariantLinkForeignKeys($schema);
         $this->addOroB2BProductShortDescriptionForeignKeys($schema);
+        $this->addOroB2BProductImageForeignKeys($schema);
+        $this->addOroB2BProductImageTypeForeignKeys($schema);
 
         $this->updateProductTable($schema);
         $this->addNoteAssociations($schema);
@@ -301,21 +308,21 @@ class OroB2BProductBundleInstaller implements
      */
     protected function addAttachmentAssociations(Schema $schema)
     {
-        $this->attachmentExtension->addImageRelation(
-            $schema,
-            self::PRODUCT_TABLE_NAME,
-            'image',
-            [
-                'importexport' => ['excluded' => true]
-            ],
-            self::MAX_PRODUCT_IMAGE_SIZE_IN_MB
-        );
-
         $this->attachmentExtension->addAttachmentAssociation(
             $schema,
             self::PRODUCT_TABLE_NAME,
             [],
             self::MAX_PRODUCT_ATTACHMENT_SIZE_IN_MB
+        );
+
+        $this->attachmentExtension->addImageRelation(
+            $schema,
+            self::PRODUCT_IMAGE_TABLE_NAME,
+            'image',
+            [
+                'importexport' => ['excluded' => true]
+            ],
+            self::MAX_PRODUCT_IMAGE_SIZE_IN_MB
         );
     }
 
@@ -381,6 +388,57 @@ class OroB2BProductBundleInstaller implements
             ['short_description_id'],
             ['id'],
             ['onDelete' => 'CASCADE', 'onUpdate' => null]
+        );
+    }
+
+    /**
+     * @param Schema $schema
+     */
+    protected function createOroB2BProductImageTable(Schema $schema)
+    {
+        $table = $schema->createTable(self::PRODUCT_IMAGE_TABLE_NAME);
+        $table->addColumn('id', 'integer', ['autoincrement' => true]);
+        $table->addColumn('product_id', 'integer', ['notnull' => true]);
+        $table->setPrimaryKey(['id']);
+    }
+
+    /**
+     * @param Schema $schema
+     */
+    protected function createOroB2BProductImageTypeTable(Schema $schema)
+    {
+        $table = $schema->createTable(self::PRODUCT_IMAGE_TYPE_TABLE_NAME);
+        $table->addColumn('id', 'integer', ['autoincrement' => true]);
+        $table->addColumn('product_image_id', 'integer');
+        $table->addColumn('type', 'string', ['length' => 255]);
+        $table->setPrimaryKey(['id']);
+    }
+
+    /**
+     * @param Schema $schema
+     */
+    protected function addOroB2BProductImageForeignKeys(Schema $schema)
+    {
+        $table = $schema->getTable(self::PRODUCT_IMAGE_TABLE_NAME);
+        $table->addForeignKeyConstraint(
+            $schema->getTable(self::PRODUCT_TABLE_NAME),
+            ['product_id'],
+            ['id'],
+            ['onUpdate' => null, 'onDelete' => 'CASCADE']
+        );
+    }
+
+    /**
+     * @param Schema $schema
+     */
+    protected function addOroB2BProductImageTypeForeignKeys(Schema $schema)
+    {
+        $table = $schema->getTable(self::PRODUCT_IMAGE_TYPE_TABLE_NAME);
+        $table->addForeignKeyConstraint(
+            $schema->getTable(self::PRODUCT_IMAGE_TABLE_NAME),
+            ['product_image_id'],
+            ['id'],
+            ['onDelete' => null, 'onUpdate' => null]
         );
     }
 }
