@@ -2,7 +2,9 @@
 
 namespace OroB2B\Bundle\ProductBundle\Tests\Functional\Controller;
 
+use Symfony\Component\DomCrawler\Crawler;
 use Symfony\Component\DomCrawler\Form;
+use Symfony\Component\HttpFoundation\File\UploadedFile;
 
 use Oro\Bundle\LocaleBundle\Entity\Localization;
 use Oro\Bundle\LocaleBundle\Entity\LocalizedFallbackValue;
@@ -49,6 +51,23 @@ class ProductControllerTest extends WebTestCase
     const CATEGORY_ID = 1;
     const CATEGORY_NAME = 'Master Catalog';
 
+    const FIRST_IMAGE_FILENAME = 'image1.gif';
+    const SECOND_IMAGE_FILENAME = 'image2.gif';
+
+    const IMAGES_VIEW_BODY_SELECTOR = 'div.image-collection table tbody tr';
+    const IMAGES_VIEW_HEAD_SELECTOR = 'div.image-collection table thead tr th';
+    const IMAGE_TYPE_CHECKED_TAG = 'i';
+    const IMAGE_TYPE_CHECKED_CLASS = 'icon-check';
+    const IMAGE_FILENAME_ATTR = 'title';
+
+    /**
+     * @var array
+     */
+    private static $expectedProductImageMatrixHeaders = ['File', 'Main', 'Listing', 'Additional'];
+
+    /**
+     * {@inheritdoc}
+     */
     protected function setUp()
     {
         $this->initClient([], $this->generateBasicAuthHeader());
@@ -101,8 +120,20 @@ class ProductControllerTest extends WebTestCase
             'sell' => true,
         ];
 
+        $formValues['orob2b_product']['images'][] = [
+            'main' => 1,
+            'listing' => 1,
+            'additional' => 1
+        ];
+
+        $filesData['orob2b_product']['images'][] = [
+            'image' => [
+                'file' => $this->createUploadedFile(self::FIRST_IMAGE_FILENAME)
+            ]
+        ];
+
         $this->client->followRedirects(true);
-        $crawler = $this->client->request($form->getMethod(), $form->getUri(), $formValues);
+        $crawler = $this->client->request($form->getMethod(), $form->getUri(), $formValues, $filesData);
 
         $result = $this->client->getResponse();
         $this->assertHtmlResponseStatusCodeEquals($result, 200);
@@ -113,9 +144,18 @@ class ProductControllerTest extends WebTestCase
         $this->assertContains(self::INVENTORY_STATUS, $html);
         $this->assertContains(self::STATUS, $html);
         $this->assertContains(self::FIRST_UNIT_CODE, $html);
+
+        $expectedProductImageMatrix = [
+            self::$expectedProductImageMatrixHeaders,
+            [self::FIRST_IMAGE_FILENAME, 1, 1, 1],
+        ];
+
+        $this->assertEquals($expectedProductImageMatrix, $this->parseProductImages($crawler));
     }
 
     /**
+     * @SuppressWarnings(PHPMD.ExcessiveMethodLength)
+     *
      * @depends testCreate
      * @return int
      */
@@ -166,12 +206,33 @@ class ProductControllerTest extends WebTestCase
                         'localizations' => [$localization->getId() => ['fallback' => FallbackType::SYSTEM]],
                     ],
                     'ids' => [$localization->getId() => $localizedName->getId()],
+                ],
+                'images' => [
+                    0 => [
+                        'main' => 1,
+                        'listing' => 1
+                    ],
+                    1 => [
+                        'additional' => 1
+                    ]
                 ]
             ],
         ];
 
+        $filesData = [
+            'orob2b_product' => [
+                'images' => [
+                    1 => [
+                        'image' => [
+                            'file' => $this->createUploadedFile(self::SECOND_IMAGE_FILENAME)
+                        ]
+                    ],
+                ]
+            ]
+        ];
+
         $this->client->followRedirects(true);
-        $this->client->request($form->getMethod(), $form->getUri(), $submittedData);
+        $this->client->request($form->getMethod(), $form->getUri(), $submittedData, $filesData);
         $result = $this->client->getResponse();
         $this->assertHtmlResponseStatusCodeEquals($result, 200);
 
@@ -229,6 +290,14 @@ class ProductControllerTest extends WebTestCase
         $this->assertContains(self::UPDATED_STATUS, $html);
         $this->assertProductPrecision($id, self::SECOND_UNIT_CODE, self::SECOND_UNIT_PRECISION);
         $this->assertProductPrecision($id, self::THIRD_UNIT_CODE, self::THIRD_UNIT_PRECISION);
+
+        $expectedProductImageMatrix = [
+            self::$expectedProductImageMatrixHeaders,
+            [self::FIRST_IMAGE_FILENAME, 1, 1, 0],
+            [self::SECOND_IMAGE_FILENAME, 0, 0, 1]
+        ];
+
+        $this->assertEquals($expectedProductImageMatrix, $this->parseProductImages($crawler));
     }
 
     /**
@@ -266,6 +335,14 @@ class ProductControllerTest extends WebTestCase
         );
         $this->assertContainsAdditionalUnitPrecision(self::SECOND_UNIT_FULL_NAME, self::SECOND_UNIT_PRECISION, $html);
         $this->assertContainsAdditionalUnitPrecision(self::THIRD_UNIT_FULL_NAME, self::THIRD_UNIT_PRECISION, $html);
+
+        $expectedProductImageMatrix = [
+            self::$expectedProductImageMatrixHeaders,
+            [self::FIRST_IMAGE_FILENAME, 1, 1, 0],
+            [self::SECOND_IMAGE_FILENAME, 0, 0, 1]
+        ];
+
+        $this->assertEquals($expectedProductImageMatrix, $this->parseProductImages($crawler));
 
         $product = $this->getProductDataBySku(self::FIRST_DUPLICATED_SKU);
 
@@ -320,6 +397,7 @@ class ProductControllerTest extends WebTestCase
                     ],
                     'ids' => [$localization->getId() => $localizedName->getId()],
                 ],
+                'images' => []//remove all images
             ],
         ];
 
@@ -344,6 +422,8 @@ class ProductControllerTest extends WebTestCase
         );
         $this->assertContainsAdditionalUnitPrecision(self::SECOND_UNIT_FULL_NAME, self::SECOND_UNIT_PRECISION, $html);
         $this->assertContainsAdditionalUnitPrecision(self::THIRD_UNIT_FULL_NAME, self::THIRD_UNIT_PRECISION, $html);
+
+        $this->assertEmpty($this->parseProductImages($crawler));
 
         $product = $this->getProductDataBySku(self::UPDATED_SKU);
 
@@ -536,5 +616,50 @@ class ProductControllerTest extends WebTestCase
             $expectedDefaultProductUnitPrecision,
             $formValues['orob2b_product[primaryUnitPrecision][precision]']
         );
+    }
+
+    /**
+     * @param string $fileName
+     * @return UploadedFile
+     */
+    private function createUploadedFile($fileName)
+    {
+        return new UploadedFile(__DIR__ . '/files/example.gif', $fileName);
+    }
+
+    /**
+     * @param Crawler $crawler
+     * @return array
+     */
+    private function parseProductImages(Crawler $crawler)
+    {
+        $result = [];
+
+        $children = $crawler->filter(self::IMAGES_VIEW_HEAD_SELECTOR);
+        /** @var \DOMElement $child */
+        foreach ($children as $child) {
+            $result[0][] = $child->textContent;
+        }
+
+        $crawler->filter(self::IMAGES_VIEW_BODY_SELECTOR)->each(
+            function (Crawler $node) use (&$result) {
+                $data = [];
+                $data[] = $node->filter('a')->first()->attr(self::IMAGE_FILENAME_ATTR);
+
+                /** @var \DOMElement $child */
+                foreach ($node->children()->nextAll() as $child) {
+                    $icon = $child->getElementsByTagName(self::IMAGE_TYPE_CHECKED_TAG)->item(0);
+                    $checked = false;
+                    if ($icon) {
+                        $iconClass = $icon->attributes->getNamedItem('class')->nodeValue;
+                        $checked = $iconClass == self::IMAGE_TYPE_CHECKED_CLASS;
+                    }
+                    $data[] = (int) $checked;
+                }
+                $result[] = $data;
+            }
+        );
+
+        return $result;
     }
 }
