@@ -5,11 +5,13 @@ namespace OroB2B\Bundle\ProductBundle\Tests\Unit\Form\Type;
 use Symfony\Component\Form\PreloadedExtension;
 use Symfony\Component\Form\Test\FormIntegrationTestCase;
 
+use Oro\Bundle\AttachmentBundle\Entity\File;
 use Oro\Bundle\AttachmentBundle\Form\Type\ImageType;
 use Oro\Bundle\EntityConfigBundle\Provider\ConfigProvider;
 use Oro\Bundle\FormBundle\Form\Extension\TooltipFormExtension;
 use Oro\Bundle\FormBundle\Form\Type\CollectionType as OroCollectionType;
 use Oro\Bundle\FormBundle\Form\Type\EntityIdentifierType;
+use Oro\Bundle\LayoutBundle\Provider\ImageTypeProvider;
 use Oro\Bundle\LocaleBundle\Entity\LocalizedFallbackValue;
 use Oro\Bundle\LocaleBundle\Form\Type\LocalizedFallbackValueCollectionType;
 use Oro\Bundle\LocaleBundle\Tests\Unit\Form\Type\Stub\LocalizedFallbackValueCollectionTypeStub;
@@ -22,20 +24,23 @@ use OroB2B\Bundle\ProductBundle\Entity\ProductUnitPrecision;
 use OroB2B\Bundle\ProductBundle\Form\Extension\IntegerExtension;
 use OroB2B\Bundle\ProductBundle\Form\Type\ProductCustomFieldsChoiceType;
 use OroB2B\Bundle\ProductBundle\Form\Type\ProductPrimaryUnitPrecisionType;
+use OroB2B\Bundle\ProductBundle\Form\Type\ProductImageCollectionType;
+use OroB2B\Bundle\ProductBundle\Form\Type\ProductImageType;
+use OroB2B\Bundle\ProductBundle\Form\Type\ProductStatusType;
 use OroB2B\Bundle\ProductBundle\Form\Type\ProductType;
 use OroB2B\Bundle\ProductBundle\Form\Type\ProductUnitPrecisionCollectionType;
 use OroB2B\Bundle\ProductBundle\Form\Type\ProductUnitPrecisionType;
 use OroB2B\Bundle\ProductBundle\Form\Type\ProductUnitSelectionType;
 use OroB2B\Bundle\ProductBundle\Form\Type\ProductVariantLinksType;
-use OroB2B\Bundle\ProductBundle\Form\Type\ProductStatusType;
 use OroB2B\Bundle\ProductBundle\Provider\DefaultProductUnitProvider;
+use OroB2B\Bundle\ProductBundle\Provider\ProductStatusProvider;
 use OroB2B\Bundle\ProductBundle\Rounding\RoundingServiceInterface;
-use OroB2B\Bundle\ProductBundle\Tests\Unit\Form\Type\Stub\ProductCustomFieldsChoiceTypeStub;
-use OroB2B\Bundle\ProductBundle\Tests\Unit\Form\Type\Stub\ProductUnitSelectionTypeStub;
+use OroB2B\Bundle\ProductBundle\Tests\Unit\Entity\Stub\StubProduct;
+use OroB2B\Bundle\ProductBundle\Tests\Unit\Entity\Stub\StubProductImage;
 use OroB2B\Bundle\ProductBundle\Tests\Unit\Form\Type\Stub\EnumSelectTypeStub;
 use OroB2B\Bundle\ProductBundle\Tests\Unit\Form\Type\Stub\ImageTypeStub;
-use OroB2B\Bundle\ProductBundle\Tests\Unit\Entity\Stub\StubProduct;
-use OroB2B\Bundle\ProductBundle\Provider\ProductStatusProvider;
+use OroB2B\Bundle\ProductBundle\Tests\Unit\Form\Type\Stub\ProductCustomFieldsChoiceTypeStub;
+use OroB2B\Bundle\ProductBundle\Tests\Unit\Form\Type\Stub\ProductUnitSelectionTypeStub;
 
 class ProductTypeTest extends FormIntegrationTestCase
 {
@@ -63,6 +68,11 @@ class ProductTypeTest extends FormIntegrationTestCase
     ];
 
     /**
+     * @var array
+     */
+    protected $images = [];
+
+    /**
      * {@inheritDoc}
      */
     protected function setUp()
@@ -80,6 +90,14 @@ class ProductTypeTest extends FormIntegrationTestCase
 
         $this->type = new ProductType($this->defaultProductUnitProvider, $this->roundingService);
         $this->type->setDataClass(self::DATA_CLASS);
+
+        $image1 = new StubProductImage();
+        $image1->setImage(new File());
+
+        $image2 = new StubProductImage();
+        $image2->setImage(new File());
+
+        $this->images = [$image1, $image2];
 
         parent::setUp();
     }
@@ -113,6 +131,13 @@ class ProductTypeTest extends FormIntegrationTestCase
         $translator = $this->getMockBuilder('Oro\Bundle\TranslationBundle\Translation\Translator')
             ->disableOriginalConstructor()
             ->getMock();
+        /** @var \PHPUnit_Framework_MockObject_MockObject|ImageTypeProvider $imageTypeProvider*/
+        $imageTypeProvider = $this->getMockBuilder('Oro\Bundle\LayoutBundle\Provider\ImageTypeProvider')
+            ->disableOriginalConstructor()
+            ->getMock();
+        $imageTypeProvider->expects($this->any())
+            ->method('getImageTypes')
+            ->willReturn([]);
 
         return [
             new PreloadedExtension(
@@ -137,6 +162,8 @@ class ProductTypeTest extends FormIntegrationTestCase
                     EntityIdentifierType::NAME => new StubEntityIdentifierType([]),
                     ProductVariantLinksType::NAME => new ProductVariantLinksType(),
                     ProductStatusType::NAME => new ProductStatusType(new ProductStatusProvider()),
+                    ProductImageCollectionType::NAME => new ProductImageCollectionType($imageTypeProvider),
+                    ProductImageType::NAME => new ProductImageType()
                 ],
                 [
                     'form' => [
@@ -258,6 +285,19 @@ class ProductTypeTest extends FormIntegrationTestCase
                 'expectedData'  => $this->createExpectedProductEntity(false, false, false),
                 'rounding' => false
             ],
+            'simple product with images' => [
+                'defaultData'   => $this->createDefaultProductEntity(false),
+                'submittedData' => [
+                    'sku' => 'test sku',
+                    'primaryUnitPrecision' => ['unit' => 'each', 'precision' => 0],
+                    'inventoryStatus' => Product::INVENTORY_STATUS_IN_STOCK,
+                    'visible' => 1,
+                    'status' => Product::STATUS_DISABLED,
+                    'images' => $this->images
+                ],
+                'expectedData'  => $this->createExpectedProductEntity(false, false, false),
+                'rounding' => false
+            ],
         ];
     }
 
@@ -270,7 +310,8 @@ class ProductTypeTest extends FormIntegrationTestCase
     protected function createExpectedProductEntity(
         $withProductUnitPrecision = false,
         $withNamesAndDescriptions = false,
-        $hasVariants = true
+        $hasVariants = true,
+        $hasImages = false
     ) {
         $expectedProduct = new StubProduct();
 
@@ -305,6 +346,12 @@ class ProductTypeTest extends FormIntegrationTestCase
                 ->addDescription($this->createLocalizedValue(null, 'second description'))
                 ->addShortDescription($this->createLocalizedValue(null, 'first short description'))
                 ->addShortDescription($this->createLocalizedValue(null, 'second short description'));
+        }
+
+        if ($hasImages) {
+            foreach ($this->images as $image) {
+                $expectedProduct->addImage($image);
+            }
         }
 
         return $expectedProduct->setSku('test sku');
