@@ -23,11 +23,6 @@ class ShoppingListController extends Controller
 {
     /**
      * @Route("/{id}", name="orob2b_shopping_list_frontend_view", defaults={"id" = null}, requirements={"id"="\d+"})
-     * @ParamConverter(
-     *     "shoppingList",
-     *     class="OroB2B\Bundle\ShoppingListBundle\Entity\ShoppingList",
-     *     isOptional="true",
-     *     options={"id" = "id"})
      * @Layout(vars={"title"})
      * @Acl(
      *      id="orob2b_shopping_list_frontend_view",
@@ -37,27 +32,32 @@ class ShoppingListController extends Controller
      *      group_name="commerce"
      * )
      *
-     * @param ShoppingList|null $shoppingList
+     * @param int|null $shoppingList
      *
      * @return array
      */
-    public function viewAction(ShoppingList $shoppingList = null)
+    public function viewAction($id = null)
     {
+        /** @var ShoppingListRepository $repo */
+        $repo = $this->getDoctrine()->getRepository('OroB2BShoppingListBundle:ShoppingList');
+        $shoppingList = $repo->findOneByIdWithRelations($id);
+
         if (!$shoppingList) {
             $user = $this->getUser();
             if ($user instanceof AccountUser) {
-                /** @var ShoppingListRepository $repo */
-                $repo = $this->getDoctrine()->getRepository('OroB2BShoppingListBundle:ShoppingList');
-                $shoppingList = $repo->findAvailableForAccountUser($user);
+                $shoppingList = $repo->findAvailableForAccountUser($user, true);
             }
         }
-
-        $totalWithSubtotalsAsArray = $shoppingList
-            ? $this->getTotalProcessor()->getTotalWithSubtotalsAsArray($shoppingList)
-            : [];
+        if ($shoppingList) {
+            $title = $shoppingList->getLabel();
+            $totalWithSubtotalsAsArray = $this->getTotalProcessor()->getTotalWithSubtotalsAsArray($shoppingList);
+        } else {
+            $title = null;
+            $totalWithSubtotalsAsArray = [];
+        }
 
         return [
-            'title' => $shoppingList ? $shoppingList->getLabel() : null,
+            'title' => $title,
             'data' => [
                 'entity' => $shoppingList,
                 'totals' => [
