@@ -31,7 +31,8 @@ class CartStrategyTest extends AbstractStrategyTest
             $this->fieldHelper,
             $this->databaseHelper,
             $this->chainEntityClassNameProvider,
-            $this->translator
+            $this->translator,
+            $this->newEntitiesHelper
         );
 
         $strategy->setOwnerHelper($this->defaultOwnerHelper);
