@@ -22,7 +22,8 @@ class CartWithExistingCustomerStrategyTest extends AbstractStrategyTest
             $this->fieldHelper,
             $this->databaseHelper,
             $this->chainEntityClassNameProvider,
-            $this->translator
+            $this->translator,
+            $this->newEntitiesHelper
         );
     }
 
