@@ -21,7 +21,8 @@ class GuestCustomerStrategyTest extends AbstractStrategyTest
             $this->fieldHelper,
             $this->databaseHelper,
             $this->chainEntityClassNameProvider,
-            $this->translator
+            $this->translator,
+            $this->newEntitiesHelper
         );
 
         $strategy->setOwnerHelper($this->defaultOwnerHelper);
