@@ -3,15 +3,17 @@
 namespace Oro\Bundle\AttachmentBundle\Guesser;
 
 use Symfony\Component\HttpFoundation\File\MimeType\MimeTypeGuesserInterface;
-
-use Oro\Component\PhpUtils\ArrayUtil;
+use Symfony\Component\HttpFoundation\File\UploadedFile;
+use Symfony\Component\HttpFoundation\FileBag;
 
 class MsMimeTypeGuesser implements MimeTypeGuesserInterface
 {
-    /** @var array */
+    /**
+     * @var array
+     */
     protected $typesMap = [
-        'd0cf11e0a1b11ae1' => [
-            'msg' => 'application/vnd.ms-outlook',
+        'msg' => [
+            'd0cf11e0a1b11ae1' => 'application/vnd.ms-outlook',
         ],
     ];
 
@@ -20,32 +22,59 @@ class MsMimeTypeGuesser implements MimeTypeGuesserInterface
      */
     public function guess($path)
     {
-        $file = ArrayUtil::find(
-            function (array $file) use ($path) {
-                if (!isset($file['tmp_name']['file'])) {
-                    $tmpName = array_pop($file['tmp_name']);
-                    return $tmpName['file'] === $path;
-                } else {
-                    return $file['tmp_name']['file'] === $path;
-                }
-            },
-            $_FILES
-        );
-
-        if (!$file) {
+        if (!$extension = $this->getExtensionByPath($path)) {
             return null;
         }
 
-        $extension = pathinfo($path, PATHINFO_EXTENSION);
+        if (!isset($this->typesMap[$extension])) {
+            return null;
+        }
 
         $handle = fopen($path, 'r');
         $bytes = bin2hex(fread($handle, 8));
         fclose($handle);
 
-        if (!isset($this->typesMap[$bytes][$extension])) {
+        if (!isset($this->typesMap[$extension][$bytes])) {
             return null;
         }
 
-        return $this->typesMap[$bytes][$extension];
+        return $this->typesMap[$extension][$bytes];
+    }
+
+    /**
+     * @param string $path
+     * @return string|null
+     */
+    private function getExtensionByPath($path)
+    {
+        $extension = null;
+        $uploadedFiles = $this->fetchUploadedFiles();
+
+        if (isset($uploadedFiles[$path])) {
+            $extension = pathinfo($uploadedFiles[$path], PATHINFO_EXTENSION);
+        }
+
+        return $extension;
+    }
+
+    /**
+     * @return array
+     */
+    private function fetchUploadedFiles()
+    {
+        $fileBagArray = (new FileBag($_FILES))->getIterator()->getArrayCopy();
+        $files = [];
+
+        array_walk_recursive(
+            $fileBagArray,
+            function ($item) use (&$files) {
+                /** $item UploadedFile */
+                if ($item instanceof UploadedFile) {
+                    $files[$item->getRealPath()] = $item->getClientOriginalName();
+                }
+            }
+        );
+
+        return $files;
     }
 }
