@@ -17,7 +17,6 @@ use Oro\Bundle\EmailBundle\Entity\EmailOrigin;
 use Oro\Bundle\EmailBundle\Entity\EmailUser;
 use Oro\Bundle\EmailBundle\Entity\InternalEmailOrigin;
 use Oro\Bundle\EmailBundle\Entity\Manager\EmailActivityManager;
-use Oro\Bundle\EmailBundle\Entity\Provider\EmailOwnerProvider;
 use Oro\Bundle\EmailBundle\Event\EmailBodyAdded;
 use Oro\Bundle\EmailBundle\Form\Model\Email as EmailModel;
 use Oro\Bundle\EmailBundle\Form\Model\EmailAttachment as EmailAttachmentModel;
@@ -25,11 +24,9 @@ use Oro\Bundle\EmailBundle\Model\FolderType;
 use Oro\Bundle\EmailBundle\Tools\EmailAddressHelper;
 use Oro\Bundle\EmailBundle\Tools\EmailOriginHelper;
 use Oro\Bundle\EntityBundle\ORM\DoctrineHelper;
-use Oro\Bundle\EntityConfigBundle\DependencyInjection\Utils\ServiceLink;
 use Oro\Bundle\ImapBundle\Entity\UserEmailOrigin;
 use Oro\Bundle\OrganizationBundle\Entity\OrganizationInterface;
 use Oro\Bundle\SecurityBundle\Encoder\Mcrypt;
-use Oro\Bundle\SecurityBundle\SecurityFacade;
 
 /**
  * Class Processor
@@ -53,23 +50,11 @@ class Processor
     /** @var EmailAddressHelper */
     protected $emailAddressHelper;
 
-    /** @var EmailEntityBuilder */
-    protected $emailEntityBuilder;
-
-    /** @var  EmailOwnerProvider */
-    protected $emailOwnerProvider;
-
     /** @var  EmailActivityManager */
     protected $emailActivityManager;
 
-    /** @var SecurityFacade */
-    protected $securityFacade;
-
     /** @var EventDispatcherInterface */
     protected $eventDispatcher;
-
-    /** @var array */
-    protected $origins = [];
 
     /** @var Mcrypt */
     protected $encryptor;
@@ -78,15 +63,11 @@ class Processor
     protected $emailOriginHelper;
 
     /**
-     * @SuppressWarnings(PHPMD.ExcessiveParameterList)
-     *
      * @param DoctrineHelper           $doctrineHelper
      * @param DirectMailer             $mailer
      * @param EmailAddressHelper       $emailAddressHelper
      * @param EmailEntityBuilder       $emailEntityBuilder
-     * @param EmailOwnerProvider       $emailOwnerProvider
      * @param EmailActivityManager     $emailActivityManager
-     * @param ServiceLink              $serviceLink
      * @param EventDispatcherInterface $eventDispatcher
      * @param Mcrypt                   $encryptor
      * @param EmailOriginHelper        $emailOriginHelper
@@ -96,9 +77,7 @@ class Processor
         DirectMailer $mailer,
         EmailAddressHelper $emailAddressHelper,
         EmailEntityBuilder $emailEntityBuilder,
-        EmailOwnerProvider $emailOwnerProvider,
         EmailActivityManager $emailActivityManager,
-        ServiceLink $serviceLink,
         EventDispatcherInterface $eventDispatcher,
         Mcrypt $encryptor,
         EmailOriginHelper $emailOriginHelper
@@ -107,9 +86,7 @@ class Processor
         $this->mailer               = $mailer;
         $this->emailAddressHelper   = $emailAddressHelper;
         $this->emailEntityBuilder   = $emailEntityBuilder;
-        $this->emailOwnerProvider   = $emailOwnerProvider; # can be removed to EOH
         $this->emailActivityManager = $emailActivityManager;
-        $this->securityFacade       = $serviceLink->getService();
         $this->eventDispatcher      = $eventDispatcher;
         $this->encryptor            = $encryptor;
         $this->emailOriginHelper    = $emailOriginHelper;
@@ -120,67 +97,36 @@ class Processor
      *
      * @param EmailModel $model
      * @param EmailOrigin $origin Origin to send email with
+     * @param bool $persist
      *
      * @return EmailUser
      * @throws \Swift_SwiftException
      */
-    public function process(EmailModel $model, $origin = null)
+    public function process(EmailModel $model, $origin = null, $persist = true)
     {
         $this->assertModel($model);
         $messageDate     = new \DateTime('now', new \DateTimeZone('UTC'));
         $parentMessageId = $this->getParentMessageId($model);
+        $message = $this->prepareMessage($model, $parentMessageId, $messageDate);
 
-        /** @var \Swift_Message $message */
-        $message = $this->mailer->createMessage();
-        if ($parentMessageId) {
-            $message->getHeaders()->addTextHeader('References', $parentMessageId);
-            $message->getHeaders()->addTextHeader('In-Reply-To', $parentMessageId);
-        }
-        $addresses = $this->getAddresses($model->getFrom());
-        $address = $this->emailAddressHelper->extractPureEmailAddress($model->getFrom());
-        $message->setDate($messageDate->getTimestamp());
-        $message->setFrom($addresses);
-        $message->setReplyTo($addresses);
-        $message->setReturnPath($address);
-        $message->setTo($this->getAddresses($model->getTo()));
-        $message->setCc($this->getAddresses($model->getCc()));
-        $message->setBcc($this->getAddresses($model->getBcc()));
-        $message->setSubject($model->getSubject());
-        $message->setBody($model->getBody(), $model->getType() === 'html' ? 'text/html' : 'text/plain');
-
-        $this->addAttachments($message, $model);
-        $this->processEmbeddedImages($message, $model);
-
-        $messageId = '<' . $message->generateId() . '>';
-
-        if ($origin === null) {
-            $origin = $this->getEmailOrigin($model->getFrom(), $model->getOrganization());
-        }
         $this->processSend($message, $origin);
 
-        $emailUser = $this->createEmailUser($model, $messageDate, $origin);
-        $emailUser->addFolder($this->getFolder($model->getFrom(), $origin));
-        $emailUser->getEmail()->setEmailBody(
-            $this->emailEntityBuilder->body($message->getBody(), $model->getType() === 'html', true)
-        );
-        $emailUser->getEmail()->setMessageId($messageId);
-        $emailUser->setSeen(true);
-        if ($parentMessageId) {
-            $emailUser->getEmail()->setRefs($parentMessageId);
+        $emailUser = $this->prepareEmailUser($model, $origin, $message, $messageDate, $parentMessageId);
+
+        if ($persist) {
+            // persist the email and all related entities such as folders, email addresses etc.
+            $this->emailEntityBuilder->getBatch()->persist($this->getEntityManager());
+            $this->persistAttachments($model, $emailUser->getEmail());
+
+            // associate the email with the target entity if exist
+            $contexts = $model->getContexts();
+            foreach ($contexts as $context) {
+                $this->emailActivityManager->addAssociation($emailUser->getEmail(), $context);
+            }
+
+            // flush all changes to the database
+            $this->getEntityManager()->flush();
         }
-
-        // persist the email and all related entities such as folders, email addresses etc.
-        $this->emailEntityBuilder->getBatch()->persist($this->getEntityManager());
-        $this->persistAttachments($model, $emailUser->getEmail());
-
-        // associate the email with the target entity if exist
-        $contexts = $model->getContexts();
-        foreach ($contexts as $context) {
-            $this->emailActivityManager->addAssociation($emailUser->getEmail(), $context);
-        }
-
-        // flush all changes to the database
-        $this->getEntityManager()->flush();
 
         $event = new EmailBodyAdded($emailUser->getEmail());
         $this->eventDispatcher->dispatch(EmailBodyAdded::NAME, $event);
@@ -195,8 +141,14 @@ class Processor
      *
      * @return EmailUser
      */
-    protected function createEmailUser(EmailModel $model, $messageDate, EmailOrigin $origin)
+    protected function createEmailUser(EmailModel $model, $messageDate, $origin)
     {
+        $owner = null;
+        $organization = null;
+        if ($origin) {
+            $owner = $origin->getOwner();
+            $organization = $origin->getOrganization();
+        }
         $emailUser = $this->emailEntityBuilder->emailUser(
             $model->getSubject(),
             $model->getFrom(),
@@ -207,15 +159,16 @@ class Processor
             Email::NORMAL_IMPORTANCE,
             $model->getCc(),
             $model->getBcc(),
-            $origin->getOwner(),
-            $origin->getOrganization()
+            $owner,
+            $organization
         );
-        $emailUser->setOrigin($origin);
-
-        if ($origin instanceof UserEmailOrigin) {
-            if ($origin->getMailbox() !== null) {
-                $emailUser->setOwner(null);
-                $emailUser->setMailboxOwner($origin->getMailbox());
+        if ($origin) {
+            $emailUser->setOrigin($origin);
+            if ($origin instanceof UserEmailOrigin) {
+                if ($origin->getMailbox() !== null) {
+                    $emailUser->setOwner(null);
+                    $emailUser->setMailboxOwner($origin->getMailbox());
+                }
             }
         }
 
@@ -235,11 +188,7 @@ class Processor
 
         //In case when 'useremailorigin' origin doesn't have folder, get folder from internal origin
         if (!$folder && $origin instanceof UserEmailOrigin) {
-            $originKey = InternalEmailOrigin::BAP.$email;
-            if (array_key_exists($originKey, $this->origins)) {
-                unset($this->origins[$originKey]);
-            }
-            $origin = $this->getEmailOrigin($email, null, InternalEmailOrigin::BAP, false);
+            $origin = $this->emailOriginHelper->getEmailOrigin($email, null, InternalEmailOrigin::BAP, false);
             return $origin->getFolder(FolderType::SENT);
         }
 
@@ -250,12 +199,12 @@ class Processor
      * Process send email message. In case exist custom smtp host/port use it
      *
      * @param \Swift_Message  $message
-     * @param UserEmailOrigin $emailOrigin
+     * @param EmailOrigin $emailOrigin
      * @throws \Swift_SwiftException
      */
     public function processSend($message, $emailOrigin)
     {
-        if ($emailOrigin instanceof UserEmailOrigin) {
+        if ($emailOrigin && $emailOrigin instanceof UserEmailOrigin) {
             /* Modify transport smtp settings */
             if ($emailOrigin->isSmtpConfigured()) {
                 $this->mailer->prepareSmtpTransport($emailOrigin);
@@ -378,7 +327,7 @@ class Processor
     }
 
     /**
-     * Find existing email origin entity by email string or create and persist new one.
+     * @deprecated since 1.9. Use {@see Oro\Bundle\EmailBundle\Tools\EmailOriginHelper} instead
      *
      * @param string                $email
      * @param OrganizationInterface $organization
@@ -393,23 +342,7 @@ class Processor
         $originName = InternalEmailOrigin::BAP,
         $enableUseUserEmailOrigin = true
     ) {
-        $originKey = $originName . $email;
-        if (!$organization && $this->securityFacade !== null && $this->securityFacade->getOrganization()) {
-            $organization = $this->securityFacade->getOrganization();
-        }
-        if (!array_key_exists($originKey, $this->origins)) {
-            $emailOwner = $this->emailOwnerProvider->findEmailOwner(
-                $this->getEntityManager(),
-                $this->emailAddressHelper->extractPureEmailAddress($email)
-            );
-
-            $origin = $this->emailOriginHelper
-                ->findEmailOrigin($emailOwner, $organization, $originName, $enableUseUserEmailOrigin);
-
-            $this->origins[$originKey] = $origin;
-        }
-
-        return $this->origins[$originKey];
+        return $this->emailOriginHelper->getEmailOrigin($email, $organization, $originName, $enableUseUserEmailOrigin);
     }
 
     /**
@@ -489,5 +422,68 @@ class Processor
         }
 
         return $this->em;
+    }
+
+    /**
+     * @param EmailModel $model
+     * @param string $parentMessageId
+     * @param \DateTime $messageDate
+     *
+     * @return \Swift_Message
+     */
+    protected function prepareMessage(EmailModel $model, $parentMessageId, $messageDate)
+    {
+        /** @var \Swift_Message $message */
+        $message = $this->mailer->createMessage();
+        if ($parentMessageId) {
+            $message->getHeaders()->addTextHeader('References', $parentMessageId);
+            $message->getHeaders()->addTextHeader('In-Reply-To', $parentMessageId);
+        }
+        $addresses = $this->getAddresses($model->getFrom());
+        $address = $this->emailAddressHelper->extractPureEmailAddress($model->getFrom());
+        $message->setDate($messageDate->getTimestamp());
+        $message->setFrom($addresses);
+        $message->setReplyTo($addresses);
+        $message->setReturnPath($address);
+        $message->setTo($this->getAddresses($model->getTo()));
+        $message->setCc($this->getAddresses($model->getCc()));
+        $message->setBcc($this->getAddresses($model->getBcc()));
+        $message->setSubject($model->getSubject());
+        $message->setBody($model->getBody(), $model->getType() === 'html' ? 'text/html' : 'text/plain');
+
+        $this->addAttachments($message, $model);
+        $this->processEmbeddedImages($message, $model);
+
+        return $message;
+    }
+
+    /**
+     * @param EmailModel $model
+     * @param EmailOrigin $origin
+     * @param \Swift_Message $message
+     * @param \DateTime $messageDate
+     * @param string $parentMessageId
+     *
+     * @return EmailUser
+     */
+    protected function prepareEmailUser(EmailModel $model, $origin, $message, $messageDate, $parentMessageId)
+    {
+        $messageId = '<' . $message->generateId() . '>';
+        $emailUser = $this->createEmailUser($model, $messageDate, $origin);
+        if ($origin) {
+            $emailUser->addFolder($this->getFolder($model->getFrom(), $origin));
+        }
+        $emailUser->getEmail()->setEmailBody(
+            $this->emailEntityBuilder->body($message->getBody(), $model->getType() === 'html', true)
+        );
+        $emailUser->getEmail()->setMessageId($messageId);
+        $emailUser->setSeen(true);
+        if ($parentMessageId) {
+            $emailUser->getEmail()->setRefs($parentMessageId);
+
+            return $emailUser;
+        }
+
+        return $emailUser;
     }
 }
