@@ -30,7 +30,6 @@ class SendEmail extends AbstractSendEmail
         EntityNameResolver $entityNameResolver
     ) {
         parent::__construct($contextAccessor, $emailProcessor, $emailAddressHelper, $entityNameResolver);
-
     }
 
     /**
@@ -94,7 +93,10 @@ class SendEmail extends AbstractSendEmail
         }
         $emailModel->setType($type);
 
-        $emailUser = $this->emailProcessor->process($emailModel);
+        $emailUser = $this->emailProcessor->process(
+            $emailModel,
+            $this->emailProcessor->getEmailOrigin($emailModel->getFrom(), $emailModel->getOrganization())
+        );
 
         if (array_key_exists('attribute', $this->options)) {
             $this->contextAccessor->setValue($context, $this->options['attribute'], $emailUser->getEmail());
