@@ -62,14 +62,7 @@ class EntityCreationTransformer extends EntityToIdTransformer
         if (!$value) {
             return null;
         } else {
-            // supported $value types:
-            // json encoded array ['value' => $valueForPropertyOfNewEntity]
-            // scalar types will be treated as ids
-            if (!is_scalar($value)) {
-                throw new UnexpectedTypeException($value, 'json encoded string or scalar value');
-            }
-            $data = json_decode($value, true);
-            $data = is_array($data) ? $data : [$this->property => $value];
+            $data = $this->getData($value);
             $id = $this->propertyAccessor->getValue($data, $this->propertyPath);
 
             return $id ? parent::reverseTransform($id) : $this->createNewEntity($data);
@@ -110,5 +103,26 @@ class EntityCreationTransformer extends EntityToIdTransformer
         }
 
         return $object;
+    }
+
+    /**
+     * @param array|int|string $value
+     *
+     * @return array
+     */
+    protected function getData($value)
+    {
+        // supported $value types:
+        // json encoded array ['value' => $valueForPropertyOfNewEntity]
+        // scalar types will be treated as ids
+        if (!is_scalar($value) && !is_array($value)) {
+            throw new UnexpectedTypeException($value, 'json encoded string, array or scalar value');
+        }
+        $data = is_scalar($value)
+            ? json_decode($value, true)
+            : $value;
+        $data = is_array($data) ? $data : [$this->property => $value];
+
+        return $data;
     }
 }
