@@ -3,6 +3,7 @@
 namespace Oro\Bundle\LocaleBundle\Entity;
 
 use Doctrine\Common\Collections\Collection;
+use Oro\Bundle\LocaleBundle\Model\FallbackType;
 
 trait FallbackTrait
 {
@@ -21,13 +22,26 @@ trait FallbackTrait
             }
         );
 
-        // TODO: implement with fallback
         if ($filteredValues->count() > 1) {
             //$title = $localization ? $localization->getTitle() : 'default';
             $title = $localization ? $localization->getName() : 'default';
             throw new \LogicException(sprintf('There must be only one %s title', $title));
         }
+
         $value = $filteredValues->first();
+        if ($value) {
+            switch ($value->getFallback()) {
+                case FallbackType::PARENT_LOCALIZATION:
+                    $value = $this->getLocalizedFallbackValue($values, $localization->getParentLocalization());
+                    break;
+                case FallbackType::SYSTEM:
+                    $value = $this->getLocalizedFallbackValue($values);
+                    break;
+                default:
+                    return $value;
+            }
+        }
+
         if (!$value && $localization !== null) {
             $value = $this->getLocalizedFallbackValue($values); // get default value
         }
