@@ -2,9 +2,13 @@
 
 namespace Oro\Bundle\WorkflowBundle\Validator;
 
+use Doctrine\ORM\EntityManager;
+
 use Symfony\Component\Validator\Mapping\ClassMetadata;
 use Symfony\Component\Validator\Mapping\Loader\AbstractLoader;
 
+use Oro\Bundle\EntityBundle\Tools\SafeDatabaseChecker;
+use Oro\Component\DependencyInjection\ServiceLink;
 use Oro\Bundle\WorkflowBundle\Restriction\RestrictionManager;
 use Oro\Bundle\WorkflowBundle\Model\WorkflowPermissionRegistry;
 
@@ -16,14 +20,32 @@ class WorkflowValidationLoader extends AbstractLoader
     /** @var RestrictionManager */
     protected $restrictionManager;
 
+    /** @var ServiceLink */
+    protected $emLink;
+
+    protected $dbCheck;
+
+    /** @var string[] */
+    protected $requiredTables = [
+        'oro_workflow_entity_acl',
+        'oro_workflow_entity_acl_ident',
+        'oro_workflow_restriction',
+        'oro_workflow_restriction_ident'
+    ];
+
     /**
      * @param WorkflowPermissionRegistry $permissionRegistry
      * @param RestrictionManager         $restrictionManager
+     * @param ServiceLink                $emLink A link to the EntityManager
      */
-    public function __construct(WorkflowPermissionRegistry $permissionRegistry, RestrictionManager $restrictionManager)
-    {
+    public function __construct(
+        WorkflowPermissionRegistry $permissionRegistry,
+        RestrictionManager $restrictionManager,
+        ServiceLink $emLink
+    ) {
         $this->permissionRegistry = $permissionRegistry;
         $this->restrictionManager = $restrictionManager;
+        $this->emLink             = $emLink;
     }
 
     /**
@@ -31,6 +53,10 @@ class WorkflowValidationLoader extends AbstractLoader
      */
     public function loadClassMetadata(ClassMetadata $metadata)
     {
+        if (!$this->checkDatabase()) {
+            return false;
+        }
+
         $class = $metadata->getClassName();
 
         if ($this->permissionRegistry->supportsClass($class, false) ||
@@ -44,5 +70,30 @@ class WorkflowValidationLoader extends AbstractLoader
         }
 
         return false;
+    }
+
+    /**
+     * @return bool
+     */
+    protected function checkDatabase()
+    {
+        if (null !== $this->dbCheck) {
+            return $this->dbCheck;
+        }
+
+        $this->dbCheck = SafeDatabaseChecker::tablesExist(
+            $this->getEntityManager()->getConnection(),
+            $this->requiredTables
+        );
+
+        return $this->dbCheck;
+    }
+
+    /**
+     * @return EntityManager
+     */
+    protected function getEntityManager()
+    {
+        return $this->emLink->getService();
     }
 }
